move modules to ui directory
