+++ src/lib/onboard/consentful/Specific.svelte
@@ -1,9 +1,9 @@
 <script lang="ts">
 	import type {Onboard_Data} from '../onboard';
-	import Markup from '$lib/Markup.svelte';
-	import Checkbox from '$lib/Checkbox.svelte';
-	import Tag from '$lib/Tag.svelte';
-	import {arrow_right} from '$lib/icons';
+	import Markup from '$lib/ui/Markup.svelte';
+	import Checkbox from '$lib/ui/Checkbox.svelte';
+	import Tag from '$lib/ui/Tag.svelte';
+	import {arrow_right} from '$lib/ui/icons';
 
 	export let data: Onboard_Data;
 	export let done: () => void;
